# Make SnapshotSubset() faster