--- a/src/rust/engine/fs/src/lib.rs
+++ b/src/rust/engine/fs/src/lib.rs
@@ -33,7 +33,10 @@ mod glob_matching_tests;
 #[cfg(test)]
 mod posixfs_tests;
 
-pub use crate::glob_matching::{GlobMatching, PreparedPathGlobs};
+pub use crate::glob_matching::{
+  ExpandablePathGlobs, GlobMatching, PathGlob, PreparedPathGlobs, DOUBLE_STAR_GLOB,
+  SINGLE_STAR_GLOB,
+};
 
 use std::cmp::min;
 use std::io::{self, Read};
@@ -198,7 +201,7 @@ impl GitignoreStyleExcludes {
     self.is_ignored_path(stat.path(), is_dir)
   }
 
-  fn is_ignored_path(&self, path: &Path, is_dir: bool) -> bool {
+  pub fn is_ignored_path(&self, path: &Path, is_dir: bool) -> bool {
     match self.gitignore.matched(path, is_dir) {
       ::ignore::Match::None | ::ignore::Match::Whitelist(_) => false,
       ::ignore::Match::Ignore(_) => true,
@@ -213,7 +216,7 @@ impl GitignoreStyleExcludes {
   }
 }
 
-#[derive(Debug)]
+#[derive(Debug, Clone)]
 pub enum StrictGlobMatching {
   // NB: the Error and Warn variants store a description of the origin of the PathGlob
   // request so that we can make the error message more helpful to users when globs fail to match.
@@ -257,7 +260,7 @@ impl StrictGlobMatching {
   }
 }
 
-#[derive(Debug)]
+#[derive(Debug, Clone)]
 pub enum GlobExpansionConjunction {
   AllMatch,
   AnyMatch,