--- a/src/rust/engine/fs/store/src/snapshot_ops.rs
+++ b/src/rust/engine/fs/store/src/snapshot_ops.rs
@@ -755,7 +755,7 @@ struct PartiallyExpandedDirectoryContext {
   pub directory_promises: Vec<PathBuf>,
 }
 
-#[derive(Clone)]
+#[derive(Clone, Debug)]
 enum RestrictedPathGlob {
   Wildcard {
     wildcard: Pattern,