@cosmicexplorer 2020-05-15
Note that `merkle_trie.rs` and the additions to `snapshot.rs` are something I stole from the `upc` project I'm working on separately to expose the pants file and directory stores to external tools: https://github.com/cosmicexplorer/upc/blob/master/local/directory/src/merkle_trie.rs. The license is of course compatible, but if anyone is wondering where all that code came from, that's where.