@cosmicexplorer 2020-05-15
For reviewers, note that it is possible to verify the speedup from this PR by:
1. checking out the first commit (right now, it's `daeb39013b92e2beb96d426832edc7363da574cd`),
2. running `./build-support/bin/native/cargo bench --manifest-path=src/rust/engine/Cargo.toml --package store -- snapshot_subset`
3. checking out HEAD and running the same command.

The output of (3) should show an improvement; something like:
```
X> ./build-support/bin/native/cargo bench --manifest-path=src/rust/engine/Cargo.toml --package store -- snapshot_subset
   Compiling store v0.1.0 (/Users/dmcclanahan/tools/p1/src/rust/engine/fs/store)
    Finished bench [optimized] target(s) in 1m 03s
     Running src/rust/engine/target/release/deps/store-19b7787f9cbb94ca
running 0 tests
test result: ok. 0 passed; 0 failed; 0 ignored; 0 measured; 103 filtered out
     Running src/rust/engine/target/release/deps/store-d540c519942f9006
Gnuplot not found, using plotters backend
Benchmarking snapshot_subset/snapshot_subset: Collecting 10 samples in estimated                                                                                  snapshot_subset/snapshot_subset
                        time:   [317.30 ms 323.56 ms 329.23 ms]
                        change: [-67.082% -66.265% -65.349%] (p = 0.00 < 0.05)
                        Performance has improved.
```