@illicitonion 2020-05-15
This is a really exciting speed-up! I guess that benchmark is showing roughly the best-case improvement (lots of mostly empty deeply nested directories), right?

I have a suggestion for a different encoding which I think may be worth considering...
With this PR as it stands, I believe we store the following:
lmdb directories store: digest bytes -> protobuf bytes
lmdb leases store: digest bytes -> timestamp
lmdb expansion cache store: digest bytes -> JSON

That JSON is likely to be one or two orders of magnitude bigger than the existing protobuf bytes, and also never gets garbage collected.

Perhaps we could instead change how we serialize directories in their primary store - instead of just the straight proto bytes, we could store a binary format which includes the transitive file infos in the same blob. That would mean that we'd have the same number of LMDBs as before (so would have disk cache locality, and avoid having to manage more files and file handles), and would mean that when we garbage-collect the directory entries, it will happen to garbage collect the expansion cache entries too. By storing in binary format, we'd save noticeable disk (and could even use cute encodings if we really wanted). And it would mean that every time we deserialise a Snapshot (which we do in more places than just subsetting) we would be able to access this information cheaply, rather than needing to iteratively do those child fetches), and without needing extra code to access a second LMDB (the `Snapshot` type requires the transitive `PathStat`s, after all).

My strawman encoding suggestions (++ means concat) is:
(length of proto bytes) ++ (proto bytes) ++ (length of FileNode proto bytes ++ FileNode-with-full-path-as-name)*

But doing some kind of merkle tree to avoid repeating common prefixes would I'm sure be possible too.

An alternative could be to hide this information in an extra field on the existing `Directory` protobuf (i.e. fork the upstream protos and add a new field - protobuf is designed for this, after all), and just make sure we clear it before using it on the wire for remote execution.

WDYT?