@cosmicexplorer 2020-05-29
@stuhood:
> Regarding the merkle_trie: AFAIK, directories and files are already stored as a merkle trie

Yes, there are many `From` implementations that allow `merkle_trie.rs` to be converted back and forth from the protobufs, so that that file can contain just specifically the logic for computing the resulting path stats of a snapshot.

> I'm also a bit hesitant to add this cache, because pantsd will cache this in memory

Ack! I think one thing that drove this design was that you constantly have to restart pantsd when editing the pants code, so node memoizations are dropped run-over-run. I agree that it would be good to avoid letting our persisted databases start growing without reason. I'll see if I can ensure that when redesigning this according to @illicitonion's comments.

Lastly, after @illicitonion's note above:
> Perhaps we could instead change how we serialize directories in their primary store - instead of just the straight proto bytes, we could store a binary format which includes the transitive file infos in the same blob.

This is wonderful, I'm going to start investigating this now.