@illicitonion 2020-05-29
> > I'm also a bit hesitant to add this cache, because pantsd will cache this in memory
> 
> Ack! I think one thing that drove this design was that you constantly have to restart pantsd when editing the pants code, so node memoizations are dropped run-over-run.

FWIW, if this is the real root cause problem, this could be interesting to fix on its own. In theory unless you're editing the core engine itself, we should be able to do more granular invalidations... In particular if you're only editing `@rule`s, we should be able to do incredibly granular invalidations.