@cosmicexplorer 2020-05-29
> I guess that benchmark is showing roughly the best-case improvement (lots of mostly empty deeply nested directories), right?

Yes, I believe that's correct!

> In particular if you're only editing @rules, we should be able to do incredibly granular invalidations.

This is an extremely interesting idea and I'm going to make an issue with it right now so we don't lose it!