@cosmicexplorer 2020-05-29
> An alternative could be to hide this information in an extra field on the existing Directory protobuf (i.e. fork the upstream protos and add a new field - protobuf is designed for this, after all), and just make sure we clear it before using it on the wire for remote execution.

I'm really feeling this approach, and will investigate this now. If we're seeing an optimization from this, it seems reasonable that others might too & we could propose it to be an actual change to the remexec CAS API. That to me seems like it would be the ideal end result of this work.